Optimise build script and turn on higher warning level

+++ game/code/game.cpp
@@ -28,8 +28,8 @@ RenderWeirdGradient(game_offscreen_buffer *Buffer, int BlueOffset, int GreenOffs
         uint32_t *Pixel = (uint32_t *)Row;
         for(int X = 0; X < Buffer->Width; ++X)
         {
-            uint8_t Blue = (X + BlueOffset);
-            uint8_t Green = (Y + GreenOffset);
+            uint8_t Blue = (uint8_t)(X + BlueOffset);
+            uint8_t Green = (uint8_t)(Y + GreenOffset);
 
             *Pixel++ = ((Green << 8) | Blue);
         }
@@ -66,7 +66,7 @@ GameUpdateAndRender(game_memory *Memory, game_input *Input, game_offscreen_buffe
     if(Input0->IsAnalogue)
     {
         // Use analogue movement tuning
-        GameState->BlueOffset += (int)4.0f*(Input0->EndX);
+        GameState->BlueOffset += (int)(4.0f*Input0->EndX);
         GameState->ToneHz = 240 + (int)(128.0f*(Input0->EndY));
 
     }
